+++ src/main/java/com/movie/entity/Order.java
@@ -16,7 +16,7 @@ public class Order {
     @TableId("id")
     private String orderId;  // 商户订单号, 主键
     private Long userId;  // 用户id, （逻辑外键，关联user.id）
-    private BigDecimal amount;  // 支付金额 //FIXME 注意：支付宝支付金额严格限定保留两位小数
+    private Double amount;  // 支付金额 //FIXME 注意：支付宝支付金额严格限定保留两位小数
     private Integer paymentStatusId;  // 支付状态id, （逻辑外键，关联payment_status.id
     private LocalDateTime createTime;  // 创建订单时间
     private LocalDateTime payTime;  // 支付时间
